Component 1 FInal Build
This is the final build of my Component 1, which prints the instructions to the program in order to display to the user. This build has had decomposition screenshots taken of it, marking its status as complete. The next program that I will now work on is Component 2, which works on "Checking Functions".