Component 4 Initial Build
This is the inital build for my Component 4, which deals with the equations associated with various shapes. Also in this Component, the Instructions Function includes a disclaimer advising users against using the program for unintended purposes, in order to address the use of the program for cheating by students. This Component will continue to be built on over the coming days.