Component 5 Working Build
This is a working build of Component 5, which now asks the user for a shape type, followed by the request for only one measurement by the program. While this program is currently functional, it does not fulfil the reqired functionality of my Component 5. This component will continued to be worked on over the course of this Assessment Period.