Final Component 7 Build
This is the final build of my Component 7, which loops the rounds of the Calculator in order to allow for multiple measurements to be given to the user. This Comopnent will now lead on to my Component 8, which compiles all of the user's answers to their selected equations throughout the program.