Component 8 Initial Build
This is the initial build of my Component 8, which provides the user with relevant statistics regarding each Round in their use of the program. This Component now has an issue with the 'round_statistics' list in the program, which combines all of the program's data into one list. This issue will most likely be resolved in the next build of this Component.