Assembled Outcome with Components 6 and 7 added
This is the build of the Assembled Outcome with Components 6 and 7 both added to the Assembled Outcome itself. Only Component 8 remains to be implemented into the program before the Assembled Outcome is, for all intents and purposes, complete, where it may go on to be Usability Tested in order to create a higher-quality outcome.