Assembled Outcome after Conclusion Formatting Added
This is a version of my Assembled Outcome that has Conclusion Formatting added to it, along with a selection of refinements in prepeartion for Usability Testing. Usability Testing will most likely be carried out on a similar version of this program, in order to complete the program in a reasonable amount of time.