Post Usability Testing Outcome with formatting, instructructions and efficiency improved
In this build of the PUTO, many aspects of the program have been improved and/or refined. The first example of this is the improved formatting of the statistics that the user recieves. The second is the improvement to the program's instructions, which have now been mostly overhauled to allow for a closer representation of the program's funcitonality. Finally, the program's efficiency has been improved, as one function is now responsible for the majority of the list appending carried out in the program, as opposed to the repeating nature of previous versions of the PUTO.