Final Post Usability Testing Outcome Build
This is the final build of the Post Usability Testing Outcome of my Geometry Calculator. This build alters the order in which some variables are defined, while also adding a selection of Comments and Subheadings, and overall fixes small inconsistencies in the cose of the program as a whole. This build of the Post Usability Testing Outcome is the one that has been used in the decomposition proof section for the program and is the final planned code version for the Geometry Calculator as a whole. This concludes the programming portion of the Geometry Calculator Assessment.